PYIC-8913: Move common tests to correct folder
